Added compressed 128x128 PNGs to the repo. Added material design font icon names to the repo in form of a JSON dictionary. - and the script scipts/webfont_icon_name_list/ to create this.
Restructured the essential main archive. It does by default not contain any SVGs anymore to guarantee this variant works on all systems, even where on cairo is present and to reduce the overall base size of SciStag. Therefore it now contains pre-rendered PNGs of all emojis and country flags in 128x128 resolution. This should work well for all standard cases and reduced the installation of SciStag to around 14 MB.

In addition to this an optional vector emoji archive can be optionally installed later from within SciStag for projects such as presentations where the large variants might be required and installation size is not that crucial.